requirements.txt
streamlit
pandas
plotly>=6.1.1
fpdf2
kaleido
psycopg2-binary
python-dotenv